requirements.txt
lifelines==0.27.8
lightning_flash==0.8.2
matplotlib==3.7.3
nibabel==3.2.2
numpy==1.23.5
pandas==1.5.3
Pillow==10.2.0
pycox==0.2.3
pytorch_warmup==0.1.1
PyYAML==6.0.1
scikit_learn==1.3.2
scikit_survival==0.22.2
scipy==1.12.0
seaborn==0.13.2
skimage==0.0
torch==2.1.0
torchio==0.19.3
torchlars==0.1.2
torchmetrics==0.10.3
torchvision==0.16.0
tqdm==4.66.1
